Wanted to upload tonight
Its a tool I thought of, after being hacked years ago. I like to be paranoid and turn off my internet when i'm not at the machine... leaving it connected just leaves more risk - to get hacked and that is NOT fun. So this auto disables Wi-Fi (in Windows only right now, but would like to make it work on Mac/Linux) whenever the mouse is idle for minutes/hours (whatever you set it at). The exe runs fine as-is on windows, but there are some bugs... with how it displays text and has the giant cmd window.